requirements.txt
python-docx>=0.8.11
tiktoken>=0.5.1
markdown>=3.4.3
python-dotenv>=1.0.0
openai>=1.3.5
requests>=2.31.0
